+++ CHANGES.md
@@ -163,6 +163,15 @@ To be released.
 [#935]: https://github.com/dahlia/optique/issues/935
 [#938]: https://github.com/dahlia/optique/pull/938
 
+### @optique/testing
+
+ -  Added the `@optique/testing` package with a shared `CapturedOutput` type
+    and reserved entry points for parser, runner, command discovery, and
+    subprocess testing.  [[#891], [#942]]
+
+[#891]: https://github.com/dahlia/optique/issues/891
+[#942]: https://github.com/dahlia/optique/pull/942
+
 
 Version 1.2.5
 -------------